moves discord info into profile image box
