+++ components/ProfileImageInfo.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import UserInfoImage from './UserInfoImage'
-import { UserInfo } from '../@types/user'
+import { UserInfo } from '../pages/profile/[username]'
+import Image from 'next/image'
 import styles from '../scss/profileImageInfo.module.scss'
 type ProfileImageInfoProps = {
   user: UserInfo
@@ -17,7 +18,15 @@ const ProfileImageInfo: React.FC<ProfileImageInfoProps> = ({ user }) => {
       <h2 className="text-center mt-4">
         {`${user.firstName} ${user.lastName}`}
       </h2>
+      <Image src="/assets/icon.svg" height={40} width={40} />
       <h4 className="text-center text-muted mb-4">{'@' + user.username}</h4>
+      <Image src="/assets/discordClydeLogo.svg" height={40} width={40} />
+      {/* https://discordapp.com/users/<id>/ */}
+      {user.discordUsername ? (
+        <h4 className="text-center text-muted mb-4">
+          {'@' + user.discordUsername}
+        </h4>
+      ) : null}
     </div>
   )
 }
