+++ pages/profile/[username].tsx
@@ -11,7 +11,6 @@ import {
 import { useUserInfoQuery } from '../../graphql/index'
 import ProfileLessons from '../../components/ProfileLessons'
 import ProfileImageInfo from '../../components/ProfileImageInfo'
-import ProfileDiscordInfo from '../../components/ProfileDiscordInfo'
 import ProfileSubmissions, {
   LessonChallenge
 } from '../../components/ProfileSubmissions'
@@ -23,9 +22,7 @@ export type UserInfo = {
   username: string
   firstName: string
   lastName: string
-}
-
-export type DiscordInfo = {
+  discordUserId: string
   discordUsername: string
   discordAvatarUrl: string
 }
@@ -60,11 +57,11 @@ const UserProfile: React.FC = () => {
     // 'A' stands for Anonymous, in case user did not put in full name
     username,
     firstName: fullname.split(' ')[0] || 'A',
-    lastName: fullname.split(' ')[1] || ' '
+    lastName: fullname.split(' ')[1] || ' ',
+    discordUserId: data?.userInfo?.user?.discordUserId ?? '',
+    discordUsername: data?.userInfo?.user?.discordUsername ?? '',
+    discordAvatarUrl: data?.userInfo?.user?.discordAvatarUrl ?? ''
   }
-
-  const discordUsername = _.get(data, 'userInfo.user.discordUsername', '')
-  const discordAvatarUrl = _.get(data, 'userInfo.user.discordAvatarUrl', '')
 
   //data is frozen so it can't be modified (reversed)
   //we need to reverse it before deduplication to keep only the latest submission
@@ -147,12 +144,6 @@ const UserProfile: React.FC = () => {
       <div className="row mt-4">
         <div className="mb-3 mb-md-0 col-md-4">
           <ProfileImageInfo user={userInfo} />
-          {discordUsername ? (
-            <ProfileDiscordInfo
-              username={discordUsername}
-              avatar={discordAvatarUrl}
-            />
-          ) : null}
           {profileStars.length === 0 ? null : (
             <ProfileStarComments stars={profileStars} />
           )}
